RPMsg-Lite update 06/2023
- QNX and ThreadX env. update
- RL_ENV_QUEUE_STATIC_STORAGE_SIZE macro correction
- Increase the rpmsg_lite queue size - could be 2U * (number of buffers allocated for one side/direction) for zero copy cases
- Code formatting using LLVM16
- Updated version to 5.1.0

+++ lib/rpmsg_lite/porting/environment/rpmsg_env_zephyr.c
@@ -672,7 +672,7 @@ void env_delete_queue(void *queue)
  * @return - status of function execution
  */
 
-int32_t env_put_queue(void *queue, void *msg, uint32_t timeout_ms)
+int32_t env_put_queue(void *queue, void *msg, uintptr_t timeout_ms)
 {
     if (env_in_isr() != 0)
     {
@@ -698,7 +698,7 @@ int32_t env_put_queue(void *queue, void *msg, uint32_t timeout_ms)
  * @return - status of function execution
  */
 
-int32_t env_get_queue(void *queue, void *msg, uint32_t timeout_ms)
+int32_t env_get_queue(void *queue, void *msg, uintptr_t timeout_ms)
 {
     if (env_in_isr() != 0)
     {
